HWKAPM-497 Enable clients to use different technologies or endpoints for publish or service query tasks. Refactored REST client implementations to provide abstract base class.
