+++ client/trace-publisher-rest-client/src/main/java/org/hawkular/apm/trace/publisher/rest/client/TracePublisherRESTClient.java
@@ -18,16 +18,15 @@ package org.hawkular.apm.trace.publisher.rest.client;
 
 import java.net.HttpURLConnection;
 import java.net.URL;
-import java.util.Base64;
 import java.util.List;
 
 import org.hawkular.apm.api.logging.Logger;
 import org.hawkular.apm.api.logging.Logger.Level;
 import org.hawkular.apm.api.model.trace.Trace;
 import org.hawkular.apm.api.services.PublisherMetricHandler;
-import org.hawkular.apm.api.services.ServiceStatus;
 import org.hawkular.apm.api.services.TracePublisher;
 import org.hawkular.apm.api.utils.PropertyUtil;
+import org.hawkular.apm.client.api.rest.AbstractRESTClient;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -37,86 +36,16 @@ import com.fasterxml.jackson.databind.ObjectMapper;
  *
  * @author gbrown
  */
-public class TracePublisherRESTClient implements TracePublisher, ServiceStatus {
+public class TracePublisherRESTClient extends AbstractRESTClient implements TracePublisher {
 
     private static final Logger log = Logger.getLogger(TracePublisherRESTClient.class.getName());
 
     private static final ObjectMapper mapper = new ObjectMapper();
 
-    private static final String HAWKULAR_TENANT = "Hawkular-Tenant";
-
-    private String username = PropertyUtil.getProperty(PropertyUtil.HAWKULAR_APM_USERNAME);
-    private String password = PropertyUtil.getProperty(PropertyUtil.HAWKULAR_APM_PASSWORD);
-
-    private String authorization = null;
-
-    private String uri;
-
     private PublisherMetricHandler<Trace> handler = null;
 
-    {
-        uri = PropertyUtil.getProperty(PropertyUtil.HAWKULAR_APM_URI);
-
-        if (uri != null && !uri.isEmpty() && uri.charAt(uri.length() - 1) != '/') {
-            uri = uri + '/';
-        }
-    }
-
-    /* (non-Javadoc)
-     * @see org.hawkular.apm.api.services.ServiceStatus#isAvailable()
-     */
-    @Override
-    public boolean isAvailable() {
-        // Check URI is specified and starts with http, so either http: or https:
-        return uri != null && uri.startsWith("http");
-    }
-
-    /**
-     * @return the username
-     */
-    public String getUsername() {
-        return username;
-    }
-
-    /**
-     * @param username the username to set
-     */
-    public void setUsername(String username) {
-        this.username = username;
-
-        // Clear any previously computed authorization string
-        this.authorization = null;
-    }
-
-    /**
-     * @return the password
-     */
-    public String getPassword() {
-        return password;
-    }
-
-    /**
-     * @param password the password to set
-     */
-    public void setPassword(String password) {
-        this.password = password;
-
-        // Clear any previously computed authorization string
-        this.authorization = null;
-    }
-
-    /**
-     * @return the uri
-     */
-    public String getUri() {
-        return uri;
-    }
-
-    /**
-     * @param uri the uri to set
-     */
-    public void setUri(String uri) {
-        this.uri = uri;
+    public TracePublisherRESTClient() {
+        super(PropertyUtil.HAWKULAR_APM_URI_PUBLISHER);
     }
 
     /* (non-Javadoc)
@@ -133,7 +62,7 @@ public class TracePublisherRESTClient implements TracePublisher, ServiceStatus {
     @Override
     public void publish(String tenantId, List<Trace> traces) throws Exception {
 
-        URL url = new URL(uri + "hawkular/apm/fragments");
+        URL url = new URL(getUri() + "hawkular/apm/fragments");
 
         if (log.isLoggable(Level.FINEST)) {
             log.finest("Publish traces [tenant=" + tenantId + "][url=" + url + "]: " + traces);
@@ -198,34 +127,6 @@ public class TracePublisherRESTClient implements TracePublisher, ServiceStatus {
     public void retry(String tenantId, List<Trace> items, String subscriber, int retryCount, long delay)
             throws Exception {
         throw new java.lang.UnsupportedOperationException("Cannot retry");
-    }
-
-    /**
-     * Add the header values to the supplied connection.
-     *
-     * @param connection The connection
-     * @param tenantId The optional tenant id
-     */
-    protected void addHeaders(HttpURLConnection connection, String tenantId) {
-        if (tenantId == null) {
-            // Check if default tenant provided as property
-            tenantId = PropertyUtil.getProperty(PropertyUtil.HAWKULAR_TENANT);
-        }
-
-        if (tenantId != null) {
-            connection.setRequestProperty(HAWKULAR_TENANT, tenantId);
-        }
-
-        if (authorization == null && username != null) {
-            String authString = username + ":" + password;
-            String encoded = Base64.getEncoder().encodeToString(authString.getBytes());
-
-            authorization = "Basic " + encoded;
-        }
-
-        if (authorization != null) {
-            connection.setRequestProperty("Authorization", authorization);
-        }
     }
 
     /* (non-Javadoc)
